* Use Int instead of Byte

+++ src/main/kotlin/com/honzamuller/simulator/Bus.kt
@@ -1,7 +1,7 @@
 package com.honzamuller.simulator
 
 class Bus : Component() {
-    var data: Byte = 0b0000000
+    var data: Int = 0b0000000
 
     override fun onClear() {
         data = 0b0000_0000
